Day 8 -- We Have A Halting Problem Problem
I didn't want the "are we done yet" to interfere with the natural VM,
and if I were being pure, I'd refactor so that I had an `InstrumentedInstruction`
that wrapped the base `Instruction` class.

This is supposed to be fun, after all.

Anyway, yield to a block for each instruction, so we can do stuff
that isn't base-level VM related.