Day 6 -- You light up my life, Lanternfish
Modeling each Fish doesn't scale on a single threaded machine.

But we know the outcome is the *same* for every fish at the same
timer value, so all we really need to do is move the counts from
one value to the next, modulo births. Now each tick is O(1) in
time and memory.