Day 7 -- The Horizontal Bop
I thought for sure that I'd need fancy caching, but nope, not this
time!

TIL:

	depths.group_by(&:itself).transform_values(&:count)

so much nicer than Day 6's

	depths.reduce(Hash.new(0)) { |memo, depth| memo[depth] += 1 }

I'm really digging the __END__ / DATA for the sample data sets.

Also the little bit of file name introspection is a bit meta, but
saves me trying to remember to set the right names.